Refactor GameSection out of SeaweedTemplate
