+++ src/lib/template/Seaweed/GameSection.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	export let email = "turnipxenon@gmail.com";
 
+	import "./seaweed.postcss";
 	import HeaderPengi from "$pkg/assets/temp/header-pengi.mp4";
 	import GithubIcon from "$pkg/assets/icons/github-mark.svg";
 	import { ItchLogoHotLink } from "$pkg/consts";
@@ -210,139 +211,3 @@
 		</section>
 	</Card>
 </section>
-
-<style lang="postcss">
-    main {
-        width: 100%;
-        overflow: visible;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-    }
-
-    h1 {
-        font-size: 2em;
-    }
-
-    h2 {
-        font-size: 1.5em;
-    }
-
-    .section-card {
-        padding: 3em;
-        max-width: 800px;
-    }
-
-    .game-card, .project-card {
-        width: 100vw;
-        max-width: 36em;
-        margin-bottom: 3em;
-    }
-
-    .game-card > video {
-        border-top-left-radius: 1em;
-        border-top-right-radius: 1em;
-        max-height: 24em;
-        width: 100%;
-        object-fit: cover;
-    }
-
-    #migrante-alberta, .game-card > img, .project-card > img {
-        border-top-left-radius: 1em;
-        border-top-right-radius: 1em;
-    }
-
-    #migrante-alberta {
-        width: 100%;
-        object-fit: cover;
-    }
-
-    .game-button {
-        @apply btn variant-filled-primary;
-    }
-
-    .game-button > img {
-        max-height: 1lh;
-        /* todo: consider night mode day mode */
-    }
-
-    .two-column-separated {
-        display: flex;
-        justify-content: space-between;
-    }
-
-    .more-section {
-        padding: 1em;
-    }
-
-    .greater-about-div {
-        display: flex;
-        flex-direction: column;
-        max-width: 36em;
-    }
-
-    .experience-and-about-div {
-        display: flex;
-        gap: 2em;
-        flex-wrap: wrap;
-        justify-content: center;
-        align-items: flex-start;
-    }
-
-    .game-card-body, .project-card-body {
-        padding: 1em;
-    }
-
-    .game-card-body > blockquote {
-        margin-bottom: 0.7lh;
-    }
-
-    .game-card-body > h1, .project-card-body > h2 {
-        text-align: center;
-    }
-
-    .game-link-section {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        gap: 1em;
-        margin-top: 1em;
-    }
-
-    .games-section, .projects-section {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 2em;
-        justify-content: center;
-    }
-
-    .projects-section {
-        align-items: flex-start;
-        /*justify-content: flex-start;*/
-    }
-
-    .long-btn-image {
-        max-height: 1lh;
-    }
-
-    .itch-promo {
-        padding: 4em;
-    }
-
-    #hepcat-video {
-        object-position: 0 0;
-    }
-
-    .title-card {
-        width: clamp(360px, 80vw, 800px);
-    }
-
-    [class*='qt-'] {
-        font-weight: var(--qt-font-weight);
-    }
-
-    .accordion-header {
-        margin-top: 0.25em;
-    }
-</style>
